GBE-746: WIP: pulling views out of remains.py
