+++ expo/gbe/views/__init__.py
@@ -1,0 +1,17 @@
+from down_view import DownView
+
+#tested
+from landing_page_view import LandingPageView
+from view_troupe_view import ViewTroupeView
+from edit_troupe_view import EditTroupeView
+from bid_act_view import BidActView
+from register_persona_view import RegisterPersonaView
+from edit_persona_view import EditPersonaView
+from clone_bid_view import CloneBidView
+from edit_act_view import EditActView
+from view_act_view import ViewActView
+
+
+from review_act_view import ReviewActView
+#untested
+from create_combo_view import CreateComboView
