+++ expo/gbe/views/remains.py
@@ -1,4 +1,5 @@
-from expo.gbe_logging import *
+from django.contrib.auth.decorators import login_required
+from expo.gbe_logging import log_func
 from django.db.models import Q
 from django.core.urlresolvers import reverse
 from django.core.exceptions import PermissionDenied
@@ -12,7 +13,6 @@ from django.http import (
     HttpResponseRedirect,
     Http404,
 )
-from django.contrib.auth.decorators import login_required
 from django.contrib.auth import (
     login,
     logout,
@@ -42,8 +42,8 @@ from gbe.ticketing_idd_interface import (
     performer_act_submittal_link,
 )
 import gbe_forms_text
-from ticketingfuncs import compute_submission
-from duration import Duration
+from gbe.ticketingfuncs import compute_submission
+from gbe.duration import Duration
 from scheduler.functions import (
     set_time_format,
     get_events_and_windows
@@ -58,696 +58,6 @@ from scheduler.models import (
 visible_bid_query = (Q(biddable_ptr__conference__status='upcoming') |
                      Q(biddable_ptr__conference__status='ongoing'))
 
-
-@log_func
-def down(request):
-    '''
-    Static "Site down" notice. Simply refers user to a static template
-    with a message.
-    '''
-    template = loader.get_template('down.tmpl')
-    context = RequestContext(request, {})
-    return HttpResponse(template.render(context))
-
-
-@login_required
-@log_func
-def landing_page(request, profile_id=None, historical=False):
-    historical = "historical" in request.GET.keys()
-    standard_context = {}
-    standard_context['events_list'] = Event.objects.all()[:5]
-    if (profile_id):
-        admin_profile = validate_perms(request, ('Registrar',
-                                                 'Volunteer Coordinator',
-                                                 'Act Coordinator',
-                                                 'Conference Coordinator',
-                                                 'Vendor Coordinator',
-                                                 'Ticketing - Admin'))
-        viewer_profile = get_object_or_404(Profile, pk=profile_id)
-        admin_message = "You are viewing a user's profile, not your own."
-    else:
-        viewer_profile = validate_profile(request, require=False)
-        admin_message = None
-
-    template = loader.get_template('gbe/landing_page.tmpl')
-    class_to_class_name = {Act: "Act",
-                           Class: "Class",
-                           Costume: "Costume",
-                           Volunteer: "Volunteer"}
-    class_to_view_name = {Act: 'act_review',
-                          Class: 'class_review',
-                          Costume: 'costume_review',
-                          Volunteer: 'volunteer_review'}
-
-    if viewer_profile:
-        bids_to_review = []
-        for bid in viewer_profile.bids_to_review():
-            bid_type = class_to_class_name.get(bid.__class__, "UNKNOWN")
-            view_name = class_to_view_name.get(bid.__class__, None)
-            if view_name:
-                url = reverse(view_name,
-                              urlconf='gbe.urls',
-                              args=[str(bid.id)])
-            else:
-                url = ""
-
-            bids_to_review += [{'bid': bid,
-                                'url': url,
-                                'action': "Review",
-                                'bid_type': bid_type}]
-
-        context = RequestContext(
-            request,
-            {'profile': viewer_profile,
-             'historical': historical,
-             'alerts': viewer_profile.alerts(historical),
-             'standard_context': standard_context,
-             'personae': viewer_profile.get_personae(),
-             'troupes': viewer_profile.get_troupes(),
-             'combos': viewer_profile.get_combos(),
-             'acts': viewer_profile.get_acts(historical),
-             'shows': viewer_profile.get_shows(),
-             'classes': viewer_profile.is_teaching(historical),
-             'proposed_classes': viewer_profile.proposed_classes(historical),
-             'vendors': viewer_profile.vendors(historical),
-             'volunteering': viewer_profile.get_volunteerbids(),
-             'costumes': viewer_profile.get_costumebids(historical),
-             'review_items': bids_to_review,
-             'tickets': get_purchased_tickets(viewer_profile.user_object),
-             'acceptance_states': acceptance_states,
-             'admin_message': admin_message,
-             'bookings': viewer_profile.schedule
-             })
-    else:
-        context = RequestContext(request,
-                                 {'standard_context': standard_context})
-    return HttpResponse(template.render(context))
-
-
-@login_required
-@log_func
-def register_persona(request, **kwargs):
-    page_title = 'Stage Persona'
-    view_title = 'Tell Us About Your Stage Persona'
-    submit_button = 'Save Persona'
-    profile = validate_profile(request, require=False)
-    if not profile:
-        return HttpResponseRedirect(reverse('home',
-                                            urlconf='gbe.urls'))
-    if request.method == 'POST':
-        form = PersonaForm(request.POST, request.FILES)
-        if form.is_valid():
-            performer = form.save(commit=True)
-            pid = profile.pk
-            if request.GET.get('next', None):
-                redirect_to = request.GET['next']
-            else:
-                redirect_to = reverse('home', urlconf='gbe.urls')
-            return HttpResponseRedirect(redirect_to)
-        else:
-            return render(request, 'gbe/bid.tmpl',
-                          {'forms': [form],
-                           'nodraft': submit_button,
-                           'page_title': page_title,
-                           'view_title': view_title,
-                           })
-    else:
-        form = PersonaForm(initial={'performer_profile': profile,
-                                    'contact': profile,
-                                    })
-        return render(request, 'gbe/bid.tmpl',
-                      {'forms': [form],
-                       'nodraft': submit_button,
-                       'page_title': page_title,
-                       'view_title': view_title,
-                       })
-
-
-@login_required
-@log_func
-def edit_troupe(request, troupe_id=None):
-    page_title = 'Manage Troupe'
-    view_title = 'Tell Us About Your Troupe'
-    submit_button = 'Save Troupe'
-    profile = validate_profile(request, require=False)
-    if not profile:
-        return HttpResponseRedirect(reverse('profile_update',
-                                            urlconf='gbe.urls') +
-                                    '?next=' +
-                                    reverse('troupe_create',
-                                            urlconf='gbe.urls'))
-    personae = profile.personae.all()
-    if len(personae) == 0:
-        return HttpResponseRedirect(reverse('persona_create',
-                                            urlconf='gbe.urls') +
-                                    '?next=' +
-                                    reverse('troupe_create',
-                                            urlconf='gbe.urls'))
-    if troupe_id:
-        troupe = get_object_or_404(Troupe, resourceitem_id=troupe_id)
-    else:
-        troupe = Troupe()
-
-    if (troupe_id > 0 and
-            request.user and
-            troupe.contact != request.user.profile):
-        return HttpResponseRedirect(reverse('troupe_view',
-                                            urlconf='gbe.urls',
-                                            args=[str(troupe_id)]))
-
-    if request.method == 'POST':
-        form = TroupeForm(request.POST, request.FILES, instance=troupe)
-        if form.is_valid():
-            form.save(commit=True)
-            return HttpResponseRedirect(reverse('home', urlconf='gbe.urls'))
-        else:
-            q = Profile.objects.filter(resourceitem_id=profile.resourceitem_id)
-            form.fields['contact'] = forms.ModelChoiceField(
-                queryset=q,
-                empty_label=None,
-                label=persona_labels['contact'])
-            return render(request,
-                          'gbe/bid.tmpl',
-                          {'forms': [form],
-                           'nodraft': submit_button,
-                           'page_title': page_title,
-                           'view_title': view_title,
-                           'view_header_text': troupe_header_text})
-    else:
-        form = TroupeForm(instance=troupe, initial={'contact': profile})
-        q = Profile.objects.filter(resourceitem_id=profile.resourceitem_id)
-        form.fields['contact'] = forms.ModelChoiceField(
-            queryset=q,
-            empty_label=None,
-            label=persona_labels['contact'])
-        return render(request, 'gbe/bid.tmpl',
-                      {'forms': [form],
-                       'nodraft': submit_button,
-                       'page_title': page_title,
-                       'view_title': view_title,
-                       'view_header_text': troupe_header_text})
-
-
-@login_required
-@log_func
-def view_troupe(request, troupe_id=None):
-    '''
-    Show troupes to troupe members, only contact should edit.
-    '''
-    profile = validate_profile(request, require=False)
-    if not profile:
-        return HttpResponseRedirect(reverse('profile_update',
-                                            urlconf='gbe.urls') +
-                                    '?next=' +
-                                    reverse('troupe_create',
-                                            urlconf='gbe.urls'))
-
-    troupe = get_object_or_404(Troupe, resourceitem_id=troupe_id)
-    form = TroupeForm(instance=troupe, prefix='The Troupe')
-    owner = ParticipantForm(
-        instance=profile,
-        prefix='Troupe Contact',
-        initial={'email': profile.user_object.email,
-                 'first_name': profile.user_object.first_name,
-                 'last_name': profile.user_object.last_name})
-
-    return render(request,
-                  'gbe/bid_view.tmpl',
-                  {'readonlyform': [form, owner]})
-
-
-@login_required
-@log_func
-def create_combo(request):
-    page_title = 'Manage Combo'
-    view_title = 'Who is in this Combo?'
-    submit_button = 'Save Combo'
-
-    profile = validate_profile(request, require=False)
-    if not profile:
-        return HttpResponseRedirect(reverse('profile_update',
-                                            urlconf='gbe.urls') +
-                                    '?next=' +
-                                    reverse('combo_create',
-                                            urlconf='gbe.urls'))
-    if not profile.personae.exists():
-        return HttpResponseRedirect(reverse('persona_create',
-                                            urlconf='gbe.urls') +
-                                    '?next=' +
-                                    reverse('combo_create',
-                                            urlconf='gbe.urls'))
-    if request.method == 'POST':
-        form = ComboForm(request.POST, request.FILES)
-        if form.is_valid():
-            troupe = form.save(commit=True)
-            troupe_id = troupe.pk
-            return HttpResponseRedirect(reverse('home'), urlconf='gbe.urls')
-        else:
-            return render(request, 'gbe/bid.tmpl',
-                          {'forms': [form],
-                           'nodraft': submit_button,
-                           'page_title': page_title,
-                           'view_title': view_title,
-                           'view_header_text': combo_header_text})
-    else:
-        form = ComboForm(initial={'contact': profile})
-        return render(request, 'gbe/bid.tmpl',
-                      {'forms': [form],
-                       'nodraft': submit_button,
-                       'page_title': page_title,
-                       'view_title': view_title,
-                       'view_header_text': combo_header_text})
-
-
-@login_required
-@log_func
-def edit_persona(request, persona_id):
-    '''
-    Modify an existing Persona object.
-    '''
-    page_title = 'Manage Persona'
-    view_title = 'Tell Us About Your Stage Persona'
-    submit_button = 'Save Persona'
-    profile = validate_profile(request, require=False)
-    if not profile:
-        return HttpResponseRedirect(reverse('profile', urlconf='gbe.urls'))
-    persona = get_object_or_404(Persona, resourceitem_id=persona_id)
-    if persona.performer_profile != profile:
-        raise PermissionDenied
-
-    if request.method == 'POST':
-        form = PersonaForm(request.POST,
-                           request.FILES,
-                           instance=persona)
-        if form.is_valid():
-            performer = form.save(commit=True)
-            return HttpResponseRedirect(reverse('home',
-                                                urlconf='gbe.urls'))
-        else:
-            return render(request,
-                          'gbe/bid.tmpl',
-                          {'forms': [form],
-                           'nodraft': submit_button,
-                           'page_title': page_title,
-                           'view_title': view_title,
-                           })
-    else:
-        form = PersonaForm(instance=persona)
-        return render(request,
-                      'gbe/bid.tmpl',
-                      {'forms': [form],
-                       'nodraft': submit_button,
-                       'page_title': page_title,
-                       'view_title': view_title,
-                       })
-
-
-@login_required
-@log_func
-def bid_act(request):
-    '''
-    Create a proposed Act object.
-    '''
-    page_title = 'Propose Act'
-    view_title = 'Propose an Act'
-    fee_link = performer_act_submittal_link(request.user.id)
-    form = ActEditForm(prefix='theact')
-    audioform = AudioInfoForm(prefix='audio')
-    lightingform = LightingInfoForm(prefix='lighting')
-    stageform = StageInfoForm(prefix='stage')
-
-    profile = validate_profile(request, require=False)
-    if not profile:
-        return HttpResponseRedirect(reverse('profile', urlconf='gbe.urls'))
-    personae = profile.personae.all()
-    draft_fields = Act().bid_draft_fields
-
-    if len(personae) == 0:
-        return HttpResponseRedirect(reverse('persona_create',
-                                            urlconf='gbe.urls') +
-                                    '?next=' +
-                                    reverse('act_create',
-                                            urlconf='gbe.urls'))
-    if request.method == 'POST':
-        '''
-        If this is a formal submit request, then do all the checking.
-        If this is a draft, only a few fields are needed, use a form
-        with fewer required fields (same model)
-        '''
-        if 'submit' in request.POST.keys():
-            form = ActEditForm(request.POST,
-                               prefix='theact')
-        else:
-            form = ActEditDraftForm(request.POST,
-                                    prefix='theact')
-        if form.is_valid():
-            # hack
-            conference = Conference.objects.filter(accepting_bids=True).first()
-            act = form.save(commit=False)
-            act.conference = conference
-            techinfo = TechInfo()
-            audioinfoform = AudioInfoForm(request.POST, prefix='theact')
-            techinfo.audio = audioinfoform.save()
-            stageinfoform = StageInfoForm(request.POST, prefix='theact')
-            techinfo.stage = stageinfoform.save()
-            lightinginfo = LightingInfo()
-            lightinginfo.save()
-            techinfo.lighting = lightinginfo
-            techinfo.save()
-
-            act.tech = techinfo
-            act.submitted = False
-            act.accepted = False
-            act.save()
-
-        else:
-            fields, requiredsub = Act().bid_fields
-            return render(
-                request,
-                'gbe/bid.tmpl',
-                {'forms': [form],
-                 'page_title': page_title,
-                 'view_title': view_title,
-                 'draft_fields': draft_fields,
-                 'fee_link': fee_link,
-                 'submit_fields': requiredsub}
-            )
-
-        if 'submit' in request.POST.keys():
-            '''
-            If this is a formal submit request, did they pay?
-            They can't submit w/out paying
-            '''
-            if verify_performer_app_paid(request.user.username):
-                act.submitted = True
-                act.save()
-                return HttpResponseRedirect(reverse('home',
-                                                    urlconf='gbe.urls'))
-            else:
-                page_title = 'Act Payment'
-                return render(
-                    request,
-                    'gbe/please_pay.tmpl',
-                    {'link': fee_link,
-                     'page_title': page_title}
-                )
-        else:
-            return HttpResponseRedirect(reverse('home', urlconf='gbe.urls'))
-
-    else:
-        form = ActEditForm(initial={'owner': profile,
-                                    'performer': personae[0]},
-                           prefix='theact')
-        q = Performer.objects.filter(contact=profile)
-        form.fields['performer'] = forms.ModelChoiceField(queryset=q)
-
-        return render(
-            request,
-            'gbe/bid.tmpl',
-            {'forms': [form],
-             'page_title': page_title,
-             'view_title': view_title,
-             'fee_link': fee_link,
-             'draft_fields': draft_fields}
-        )
-
-
-@login_required
-def clone_bid(request, bid_type, bid_id):
-    '''
-    "Revive" an existing bid for use in the existing conference
-    '''
-    owner = {'Act': lambda bid: bid.performer.contact,
-             'Class': lambda bid: bid.teacher.contact,
-             'Vendor': lambda bid: bid.profile}
-
-    if bid_type not in ('Act', 'Class', 'Vendor'):
-        raise Http404   # or something
-    bid = eval(bid_type).objects.get(pk=bid_id)
-    owner_profile = owner[bid_type](bid)
-    if request.user.profile != owner_profile:
-        raise PermissionDenied
-    new_bid = bid.clone()
-    return landing_page(request)
-
-
-@login_required
-@log_func
-def edit_act(request, act_id):
-    '''
-    Modify an existing Act object.
-    '''
-    page_title = 'Edit Act Proposal'
-    view_title = 'Edit Your Act Proposal'
-    fee_link = performer_act_submittal_link(request.user.id)
-    form = ActEditForm(prefix='theact')
-
-    profile = validate_profile(request, require=False)
-    if not profile:
-        return HttpResponseRedirect(reverse('profile', urlconf='gbe.urls'))
-
-    act = get_object_or_404(Act, id=act_id)
-    if act.performer.contact != profile:
-        raise Http404
-
-    audio_info = act.tech.audio
-    stage_info = act.tech.stage
-    draft_fields = Act().bid_draft_fields
-
-    if request.method == 'POST':
-        '''
-        If this is a formal submit request, then do all the checking.
-        If this is a draft, only a few fields are needed, use a form
-        with fewer required fields (same model)
-        '''
-        if 'submit' in request.POST.keys():
-            form = ActEditForm(request.POST,
-                               instance=act,
-                               prefix='theact',
-                               initial={
-                                   'track_title': audio_info.track_title,
-                                   'track_artist': audio_info.track_artist,
-                                   'track_duration': audio_info.track_duration,
-                                   'act_duration': stage_info.act_duration
-                               })
-        else:
-            form = ActEditDraftForm(
-                request.POST,
-                instance=act,
-                prefix='theact',
-                initial={
-                    'track_title': audio_info.track_title,
-                    'track_artist': audio_info.track_artist,
-                    'track_duration': audio_info.track_duration,
-                    'act_duration': stage_info.act_duration
-                })
-        audioform = AudioInfoForm(request.POST, prefix='theact',
-                                  instance=audio_info)
-        stageform = StageInfoForm(request.POST, prefix='theact',
-                                  instance=stage_info)
-
-        if all([form.is_valid(),
-                audioform.is_valid(),
-                stageform.is_valid()]):
-            tech = act.tech
-            tech.audio = audioform.save()
-            tech.stage = stageform.save()
-            tech.save()
-            form.save()
-        else:
-            fields, requiredsub = Act().bid_fields
-            return render(
-                request,
-                'gbe/bid.tmpl',
-                {'forms': [form],
-                 'page_title': page_title,
-                 'view_title': view_title,
-                 'draft_fields': draft_fields,
-                 'fee_link': fee_link,
-                 'submit_fields': requiredsub}
-            )
-
-        if 'submit' in request.POST.keys():
-            '''
-            If this is a formal submit request, did they pay?
-            They can't submit w/out paying
-            '''
-            if verify_performer_app_paid(request.user.username):
-                act.submitted = True
-                act.save()
-                return HttpResponseRedirect(reverse('home',
-                                                    urlconf='gbe.urls'))
-            else:
-                page_title = 'Act Payment'
-                return render(
-                    request,
-                    'gbe/please_pay.tmpl',
-                    {'link': fee_link,
-                     'page_title': page_title}
-                )
-        else:
-            return HttpResponseRedirect(reverse('home', urlconf='gbe.urls'))
-    else:
-        audio_info = act.tech.audio
-        stage_info = act.tech.stage
-
-        form = ActEditForm(instance=act,
-                           prefix='theact',
-                           initial={
-                               'track_title': audio_info.track_title,
-                               'track_artist': audio_info.track_artist,
-                               'track_duration': audio_info.track_duration,
-                               'act_duration': stage_info.act_duration
-                           })
-        q = Performer.objects.filter(contact=profile)
-        form.fields['performer'] = forms.ModelChoiceField(queryset=q)
-        return render(
-            request,
-            'gbe/bid.tmpl',
-            {'forms': [form],
-             'page_title': page_title,
-             'view_title': view_title,
-             'fee_link': fee_link,
-             'draft_fields': draft_fields}
-        )
-
-
-@login_required
-@log_func
-def view_act(request, act_id):
-    '''
-    Show a bid as a read-only form.
-    '''
-    act = get_object_or_404(Act, id=act_id)
-    if act.performer.contact != request.user.profile:
-        validate_perms(request, ('Act Reviewers',), require=True)
-    audio_info = act.tech.audio
-    stage_info = act.tech.stage
-    actform = ActEditForm(
-        instance=act,
-        prefix='The Act',
-        initial={'track_title': audio_info.track_title,
-                 'track_artist': audio_info.track_artist,
-                 'track_duration': audio_info.track_duration,
-                 'act_duration': stage_info.act_duration}
-    )
-
-    if Troupe.objects.filter(pk=act.performer.pk).exists():
-        instance = Troupe.objects.get(pk=act.performer.pk)
-        performer = TroupeForm(instance=instance,
-                               prefix='The Troupe')
-    else:
-        performer = PersonaForm(instance=act.performer,
-                                prefix='The Performer(s)')
-
-    return render(request, 'gbe/bid_view.tmpl',
-                  {'readonlyform': [actform, performer]})
-
-
-def _create_action_form(act):
-    actionform = BidStateChangeForm(instance=act)
-    # This requires that the show be scheduled - seems reasonable in
-    # current workflow and lets me order by date.  Also - assumes
-    # that shows are only scheduled once
-    try:
-        start = Show.objects.filter(
-            scheduler_events__resources_allocated__resource__actresource___item=act)[0]
-    except:
-        start = ""
-    q = Show.objects.filter(
-        conference=act.conference,
-        scheduler_events__isnull=False).order_by(
-            'scheduler_events__starttime')
-    actionform.fields['show'] = forms.ModelChoiceField(
-        queryset=q,
-        empty_label=None,
-        label='Pick a Show',
-        initial=start)
-    actionURL = reverse('act_changestate',
-                        urlconf='gbe.urls',
-                        args=[act.id])
-    return actionform, actionURL
-
-
-@login_required
-@log_func
-def review_act(request, act_id):
-    '''
-    Show a bid  which needs to be reviewed by the current user.
-    To show: display all information about the bid, and a standard
-    review form.
-    If user is not a reviewer, politely decline to show anything.
-    TODO: refactor to eliminate duplication between this and view_act
-    '''
-    reviewer = validate_perms(request, ('Act Reviewers', ))
-    act = get_object_or_404(
-        Act,
-        id=act_id
-    )
-    if not act.is_current:
-        return view_act(request, act_id)
-    conference, old_bid = get_conf(act)
-    audio_info = act.tech.audio
-    stage_info = act.tech.stage
-    actform = ActEditForm(instance=act,
-                          prefix='The Act',
-                          initial={
-                              'track_title': audio_info.track_title,
-                              'track_artist': audio_info.track_artist,
-                              'track_duration': audio_info.track_duration,
-                              'act_duration': stage_info.act_duration
-                          })
-    performer = PersonaForm(instance=act.performer,
-                            prefix='The Performer(s)')
-
-    if validate_perms(request, ('Act Coordinator',), require=False):
-        actionform, actionURL = _create_action_form(act)
-    else:
-            actionform = False
-            actionURL = False
-
-    '''
-    if user has previously reviewed the act, provide their review for update
-    '''
-    try:
-        bid_eval = BidEvaluation.objects.filter(
-            bid_id=act_id,
-            evaluator_id=reviewer.resourceitem_id)[0]
-    except:
-        bid_eval = BidEvaluation(evaluator=reviewer, bid=act)
-
-    # show act info and inputs for review
-    if request.method == 'POST':
-        form = BidEvaluationForm(request.POST, instance=bid_eval)
-        if form.is_valid():
-            evaluation = form.save(commit=False)
-            evaluation.evaluator = reviewer
-            evaluation.bid = act
-            evaluation.save()
-            return HttpResponseRedirect(reverse('act_review_list',
-                                                urlconf='gbe.urls'))
-        else:
-            return render(request, 'gbe/bid_review.tmpl',
-                          {'readonlyform': [actform, performer],
-                           'reviewer': reviewer,
-                           'form': form,
-                           'actionform': actionform,
-                           'actionURL': actionURL,
-                           'conference': conference,
-                           'old_bid': old_bid,
-                           })
-    else:
-        form = BidEvaluationForm(instance=bid_eval)
-        return render(request,
-                      'gbe/bid_review.tmpl',
-                      {'readonlyform': [actform, performer],
-                       'reviewer': reviewer,
-                       'form': form,
-                       'actionform': actionform,
-                       'actionURL': actionURL,
-                       'conference': conference,
-                       'old_bid': old_bid,
-                       })
 
 
 @login_required
