+++ expo/tests/gbe/test_clone_bid.py
@@ -5,7 +5,7 @@ import nose.tools as nt
 from unittest import TestCase
 from django.test.client import RequestFactory
 from django.test import Client
-from gbe.views import clone_bid
+
 
 from tests.factories.gbe_factories import (
     ActFactory,
@@ -16,13 +16,15 @@ from gbe.models import (
     Act,
     Conference,
 )
-
+from tests.functions.gbe_functions import login_as
 
 class TestCloneBid(TestCase):
+    view_name = 'clone_bid'
+
     def setUp(self):
         self.factory = RequestFactory()
         self.client = Client()
-        self.urlstring = 'clone_bid'
+
 
     def test_clone_bid_succeed(self):
         Conference.objects.all().delete()
@@ -33,18 +35,18 @@ class TestCloneBid(TestCase):
         bid = ActFactory(conference=old_conference)
         Act.objects.filter(title=bid.title,
                            conference=current_conference).delete()
-        request = self.factory.get(self.urlstring,
-                                   urlconf="gbe.urls",
-                                   args=('Act', bid.id))
-        request.session = {'cms_admin_site': 1}
-        request.user = bid.performer.contact.user_object
 
-        response = clone_bid(request, 'Act', bid.id)
+        url = reverse(self.view_name,
+                      urlconf="gbe.urls",
+                      kwargs={'bid_type': 'Act',
+                              'bid_id': bid.id})
+        login_as(bid.performer.contact, self)
+
+        response = self.client.get(url)
         nt.assert_true(Act.objects.filter(
             title=bid.title,
             conference=current_conference).exists())
 
-    @nt.raises(Http404)
     def test_clone_bid_bad_bid_type(self):
         Conference.objects.all().delete()
         old_conference = ConferenceFactory(status="completed",
@@ -54,14 +56,14 @@ class TestCloneBid(TestCase):
         bid = ActFactory(conference=old_conference)
         Act.objects.filter(title=bid.title,
                            conference=current_conference).delete()
-        request = self.factory.get(self.urlstring,
-                                   urlconf="gbe.urls",
-                                   args=('Act', bid.id))
-        request.session = {'cms_admin_site': 1}
-        request.user = bid.performer.contact.user_object
-        response = clone_bid(request, 'Show', bid.id)
+        url = reverse(self.view_name,
+                      urlconf="gbe.urls",
+                      kwargs={'bid_type': 'Steakknife',
+                              'bid_id': bid.id})
+        login_as(bid.performer.contact, self)
+        response = self.client.get(url)
+        nt.assert_equal(response.status_code, 404)
 
-    @nt.raises(PermissionDenied)
     def test_clone_bid_wrong_user(self):
         Conference.objects.all().delete()
         old_conference = ConferenceFactory(status="completed",
@@ -71,9 +73,10 @@ class TestCloneBid(TestCase):
         bid = ActFactory(conference=old_conference)
         Act.objects.filter(title=bid.title,
                            conference=current_conference).delete()
-        request = self.factory.get(self.urlstring,
-                                   urlconf="gbe.urls",
-                                   args=('Act', bid.id))
-        request.session = {'cms_admin_site': 1}
-        request.user = ProfileFactory().user_object
-        response = clone_bid(request, 'Act', bid.id)
+        url = reverse(self.view_name,
+                      urlconf="gbe.urls",
+                      kwargs={'bid_type': 'Act',
+                              'bid_id': bid.id})
+        login_as(ProfileFactory(), self)
+        response = self.client.get(url)
+        nt.assert_equal(response.status_code, 403)
