+++ expo/tests/gbe/test_edit_troupe.py
@@ -1,9 +1,8 @@
-import gbe.models as conf
 import nose.tools as nt
 from unittest import TestCase
+from django.core.urlresolvers import reverse
 from django.test.client import RequestFactory
 from django.test import Client
-from gbe.views import edit_troupe
 from tests.factories.gbe_factories import (
     PersonaFactory,
     ProfileFactory,
@@ -19,35 +18,42 @@ from tests.functions.gbe_functions import (
 # create combos but we can't edit them. This will have to be looked at.
 
 
-class TestEditTroupe(TestCase):
+class TestCreateTroupe(TestCase):
     '''Tests for edit_troupe view'''
+
+    view_name = 'troupe_create'
+
     def setUp(self):
         self.factory = RequestFactory()
         self.client = Client()
         self.troupe_string = 'Tell Us About Your Troupe'
 
-    def test_create_troupe(self):
+    def test_create_troupe_no_performer(self):
         '''edit_troupe view, create flow
         '''
-        contact = PersonaFactory()
-        request = self.factory.get('/troupe/create/')
-        request.user = UserFactory()
-        response = edit_troupe(request)
+        login_as(UserFactory(), self)
+        url = reverse(self.view_name, urlconf='gbe.urls')
+        response = self.client.get(url)
         self.assertEqual(response.status_code, 302)
+        expected_loc = 'http://testserver/update_profile?next=/troupe/create'
+        nt.assert_equal(location(response), expected_loc)
 
-        user = UserFactory()
-        login_as(user, self)
-        request.user = user
-        response = edit_troupe(request)
-        nt.assert_equal(response.status_code, 302)
-        nt.assert_equal(location(response),
-                        '/update_profile?next=/troupe/create')
-        request.user = contact.performer_profile.user_object
-        request.session = {'cms_admin_site': 1}
+
+    def test_create_troupe_performer_exists(self):
+        contact = PersonaFactory()
         login_as(contact.performer_profile, self)
-        response = edit_troupe(request)
+        url = reverse(self.view_name, urlconf='gbe.urls')
+        response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
         self.assertTrue(self.troupe_string in response.content)
+
+
+class TestEditTroupe(TestCase):
+    view_name = 'troupe_edit'
+
+    def setUp(self):
+        self.factory = RequestFactory()
+        self.client = Client()
 
     def test_edit_troupe(self):
         '''edit_troupe view, edit flow success
@@ -55,33 +61,22 @@ class TestEditTroupe(TestCase):
         persona = PersonaFactory()
         contact = persona.performer_profile
         troupe = TroupeFactory(contact=contact)
-        request = self.factory.get('/troupe/edit/%d' % troupe.pk)
-        request.user = contact.profile.user_object
-        request.session = {'cms_admin_site': 1}
-        self.client.logout()
-        response = edit_troupe(request)
+        url = reverse(self.view_name,
+                      args=[troupe.pk],
+                      urlconf='gbe.urls')
+        login_as(contact.profile, self)
+        response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
 
     def test_no_persona(self):
         profile = ProfileFactory()
         troupe = TroupeFactory()
+        url = reverse(self.view_name,
+                      args=[troupe.pk],
+                      urlconf='gbe.urls')
         request = self.factory.get('/troupe/edit/%d' % troupe.pk)
-        request.user = profile.user_object
         login_as(profile, self)
-        request.session = {'cms_admin_site': 1}
-        response = edit_troupe(request)
-        nt.assert_equal('/performer/create?next=/troupe/create',
-                        location(response))
-        self.assertEqual(response.status_code, 302)
-
-    def test_no_persona(self):
-        profile = ProfileFactory()
-        troupe = TroupeFactory()
-        request = self.factory.get('/troupe/edit/%d' % troupe.pk)
-        request.user = profile.user_object
-        login_as(profile, self)
-        request.session = {'cms_admin_site': 1}
-        response = edit_troupe(request)
-        nt.assert_equal('/performer/create?next=/troupe/create',
-                        location(response))
+        response = self.client.get(url)
+        expected_loc = 'http://testserver/performer/create?next=/troupe/create'
+        nt.assert_equal(expected_loc, location(response))
         self.assertEqual(response.status_code, 302)
