+++ expo/tests/gbe/test_landing_page.py
@@ -1,4 +1,3 @@
-import gbe.models as conf
 import nose.tools as nt
 from unittest import TestCase
 from datetime import datetime
@@ -6,7 +5,6 @@ import pytz
 from django.test import Client
 from django.test.client import RequestFactory
 from django.core.urlresolvers import reverse
-from gbe.views import landing_page
 from tests.factories.gbe_factories import(
     ActFactory,
     ClassFactory,
@@ -32,6 +30,8 @@ from tests.functions.gbe_functions import (
 
 class TestIndex(TestCase):
     '''Tests for index view'''
+    view_name = 'home'
+
     def setUp(self):
         self.client = Client()
         # Conference Setup
@@ -145,20 +145,17 @@ class TestIndex(TestCase):
                         args=[event.eventitem.eventitem_id]) in content)
 
     def test_no_profile(self):
-        request = self.factory.get(reverse('home', urlconf="gbe.urls"))
-        request.user = UserFactory()
-        login_as(request.user, self)
-        request.session = {'cms_admin_site': 1}
-        response = landing_page(request)
+        url = reverse('home', urlconf="gbe.urls")
+        login_as(UserFactory(), self)
+        response = self.client.get(url)
         nt.assert_true("Your Expo" in response.content)
 
     def test_landing_page_path(self):
         '''Basic test of landing_page view
         '''
-        request = self.factory.get('/')
-        request.user = self.profile.user_object
-        request.session = {'cms_admin_site': 1}
-        response = landing_page(request)
+        url = reverse('home', urlconf='gbe.urls')
+        login_as(self.profile, self)
+        response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
         content = response.content
         does_not_show_previous = (
@@ -187,11 +184,11 @@ class TestIndex(TestCase):
             self.previous_class_sched, content))
 
     def test_historical_view(self):
-        request = self.factory.get('/')
-        request.user = self.profile.user_object
-        request.GET = {'historical': 1}
-        request.session = {'cms_admin_site': 1}
-        response = landing_page(request)
+        url = reverse('home', urlconf='gbe.urls')
+        login_as(self.profile, self)
+        response = self.client.get(
+            url,
+            data={'historical': 1})
         content = response.content
         self.assertEqual(response.status_code, 200)
         shows_all_previous = (
@@ -214,75 +211,65 @@ class TestIndex(TestCase):
             self.previous_class_sched, content))
 
     def test_as_privileged_user(self):
-        '''Basic test of landing_page view
-        '''
-        request = self.factory.get('/')
         staff_profile = ProfileFactory()
         grant_privilege(staff_profile, "Ticketing - Admin")
         login_as(staff_profile, self)
-        request.user = staff_profile.user_object
-        request.session = {'cms_admin_site': 1}
-        response = landing_page(request, staff_profile.pk)
+        url = reverse('admin_landing_page', urlconf='gbe.urls',
+                      args=[staff_profile.pk])
+        response = self.client.get(url)
+
         self.assertEqual(response.status_code, 200)
-        content = response.content
-        nt.assert_true("You are viewing a" in content)
+        nt.assert_true("You are viewing a" in response.content)
 
     def test_acts_to_review(self):
-        request = self.factory.get('/')
         staff_profile = ProfileFactory(user_object__is_staff=True)
         grant_privilege(staff_profile, "Act Reviewers")
         login_as(staff_profile, self)
         act = ActFactory(submitted=True,
                          conference=self.current_conf)
-        request.session = {'cms_admin_site': 1}
-        request.user = staff_profile.user_object
-        response = landing_page(request)
+        url = reverse('home', urlconf='gbe.urls')
+        response = self.client.get(url)
         nt.assert_true(act.title in response.content)
 
     def test_classes_to_review(self):
-        request = self.factory.get('/')
         staff_profile = ProfileFactory(user_object__is_staff=True)
         grant_privilege(staff_profile, "Class Reviewers")
         login_as(staff_profile, self)
         klass = ClassFactory(submitted=True,
                              conference=self.current_conf)
-        request.session = {'cms_admin_site': 1}
-        request.user = staff_profile.user_object
-        response = landing_page(request)
+        url = reverse('home', urlconf='gbe.urls')
+        response = self.client.get(url)
         nt.assert_true(klass.title in response.content)
 
     def test_volunteers_to_review(self):
-        request = self.factory.get('/')
         staff_profile = ProfileFactory(user_object__is_staff=True)
         grant_privilege(staff_profile, "Volunteer Reviewers")
         login_as(staff_profile, self)
         volunteer = VolunteerFactory(submitted=True,
                                      conference=self.current_conf)
-        request.session = {'cms_admin_site': 1}
-        request.user = staff_profile.user_object
-        response = landing_page(request)
+
+        url = reverse('home', urlconf='gbe.urls')
+        response = self.client.get(url)
         nt.assert_true(volunteer.title in response.content)
 
     def test_vendors_to_review(self):
-        request = self.factory.get('/')
         staff_profile = ProfileFactory(user_object__is_staff=True)
         grant_privilege(staff_profile, "Vendor Reviewers")
         login_as(staff_profile, self)
         vendor = VendorFactory(submitted=True,
                                conference=self.current_conf)
-        request.session = {'cms_admin_site': 1}
-        request.user = staff_profile.user_object
-        response = landing_page(request)
+        url = reverse('home', urlconf='gbe.urls')
+        response = self.client.get(url)
+
         nt.assert_true(vendor.title in response.content)
 
     def test_costumes_to_review(self):
-        request = self.factory.get('/')
         staff_profile = ProfileFactory(user_object__is_staff=True)
         grant_privilege(staff_profile, "Costume Reviewers")
         login_as(staff_profile, self)
         costume = CostumeFactory(submitted=True,
                                  conference=self.current_conf)
-        request.session = {'cms_admin_site': 1}
-        request.user = staff_profile.user_object
-        response = landing_page(request)
+        url = reverse('home', urlconf='gbe.urls')
+        response = self.client.get(url)
+
         nt.assert_true(costume.title in response.content)
