+++ expo/tests/gbe/test_view_act.py
@@ -1,16 +1,17 @@
 import nose.tools as nt
+from django.core.urlresolvers import reverse
 from unittest import TestCase
 from django.test.client import RequestFactory
 from django.test import Client
-from gbe.views import view_act
 from tests.factories.gbe_factories import (
     PersonaFactory,
     ActFactory,
 )
-
+from tests.functions.gbe_functions import login_as
 
 class TestViewAct(TestCase):
     '''Tests for view_act view'''
+    view_name = 'act_view'
 
     def setUp(self):
         self.factory = RequestFactory()
@@ -19,10 +20,11 @@ class TestViewAct(TestCase):
 
     def test_view_act_all_well(self):
         act = ActFactory()
-        request = self.factory.get('act/view/%d' % act.pk)
-        request.user = act.performer.performer_profile.user_object
-        request.session = {'cms_admin_site': 1}
-        response = view_act(request, act.pk)
+        url = reverse(self.view_name,
+                      args=[act.pk],
+                      urlconf='gbe.urls')
+        login_as(act.performer.performer_profile, self)
+        response = self.client.get(url)
         test_string = 'Submitted proposals cannot be modified'
         nt.assert_equal(response.status_code, 200)
         nt.assert_true(test_string in response.content)
