+++ expo/tests/gbe/test_view_troupe.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 from django.core.urlresolvers import reverse
 from django.test.client import RequestFactory
 from django.test import Client
-from gbe.views import view_troupe
+from gbe.views import ViewTroupeView
 from tests.factories.gbe_factories import (
     TroupeFactory,
     PersonaFactory,
@@ -11,7 +11,6 @@ from tests.factories.gbe_factories import (
 )
 from tests.functions.gbe_functions import (
     login_as,
-    is_profile_update_page,
 )
 
 
@@ -28,25 +27,21 @@ class TestViewTroupe(TestCase):
         persona = PersonaFactory()
         contact = persona.performer_profile
         troupe = TroupeFactory(contact=contact)
-        request = self.factory.get('/troupe/view/%d' % troupe.resourceitem_id)
-        request.session = {'cms_admin_site': 1}
-        request.user = contact.profile.user_object
-        response = view_troupe(request, troupe.resourceitem_id)
+        url = reverse('troupe_view',
+                      args=[troupe.resourceitem_id],
+                      urlconf='gbe.urls')
+        login_as(contact.profile.user_object, self)
+
+        response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
 
     def test_no_profile(self):
         troupe = TroupeFactory()
         user = UserFactory()
-        request = self.client.get(reverse(
+        url = reverse(
             "troupe_view",
             args=[troupe.resourceitem_id],
-            urlconf="gbe.urls"),
-                               follow=True)
-        request.session = {'cms_admin_site': 1}
-        request.user = user
-        response = view_troupe(request,
-                               troupe.resourceitem_id)
+            urlconf='gbe.urls')
+        login_as(user, self)
+        response = self.client.get(url)
         nt.assert_equal(302, response.status_code)
-        # this is not ideal, we should be getting the redirect page, but
-        # I'm having trouble with the follow parameter.
-        # Good enough for now.
